Integrated simsi_ascore into SIMSI
Integrated the subpackage into the main.py file of simsi such that ascores are calculated for the summary files. This is only done of the mode is specified as 'ascore' in the command line input, otherwise no ascoring is performed. Also changed the parsing, such that, the outputfiles of the ascorign are located where all the other outpufiles of simsi are located.